- Methode zum Öffnen der Initial-Page nach vollständigem Aufbau des GUIs implementiert gewählte Page: "Übersicht" -Methodenaufruf in GUI
