+++ js/Questionnaire.js
@@ -2,21 +2,193 @@
 
 export default class Questionnaire
 {
-	constructor(id, className, subject, amountStudents)
+	constructor(questionnaire, questionnaireList)
 	{
-		this.id = id;
-		this.className = className;
-		this.subject = subject;
-		this.amountStudents = amountStudents;
-		this.questions = [];
-		this.currentStatus = "offline";
-		this.year = 0;
-	}
+		console.log("Aufruf von Klasse Questionnaire:");
+		console.log(questionnaire);
+
+		questionnaireList.appendChild(document.createElement("br"));
+
+		this.menuBarColor;
+		this.id;
+		this.className;
+		this.subject;
+		this.amountStudents;
+		this.questions;
+		this.currentStatus;
+		this.date;
+
+		let div = document.getElementById(questionnaire.id);
+		if (div !== null) div.remove();
+		else
+		{
+			div = document.createElement("div");
+			div.style.borderStyle = "solid";
+			div.style.borderColor = "#9eb3c7";
+			div.style.borderWidth = "1px";
+			div.style.width = "99%";
+			div.id = questionnaire.id;
 	
-	addQuestion(question)
+			let tds = div.getElementsByClassName("questionnaireHeader");
+			div.addEventListener("mouseenter", ()=>{
+				div.style.cursor = "pointer";
+				for(let td in tds)
+				{
+					if (tds[td].style != undefined)
+					{
+						tds[td].style.color = "#ffffff";
+						tds[td].style.backgroundColor = this.menuBarColor;
+						
+					}
+				}
+			});
+			div.addEventListener("mouseleave", ()=>{
+				for(let td in tds)
+				{
+					if (tds[td].style != undefined)
+					{
+						tds[td].style.color = this.menuBarColor;
+						tds[td].style.backgroundColor = "#9eb3c7";
+					}
+				}
+			});
+			
+			let table = document.createElement("table");
+			table.style.borderCollapse = "collapse";
+			table.style.tableLayout = "fixed";
+			table.style.width = "100%";
+	
+			let rowHeaders = document.createElement("tr");
+			let rowData = document.createElement("tr");
+	
+			// Dokument-Symbol hinzufügen
+			let columnSymbol = document.createElement("td");
+			columnSymbol.className = "questionnaireHeader";
+			columnSymbol.rowSpan = 2;
+			columnSymbol.style.width = "50px";
+			columnSymbol.style.fontSize = "40px";
+			columnSymbol.style.textAlign = "center";
+			columnSymbol.style.color = this.menuBarColor;
+			columnSymbol.innerHTML = " &#128462;"; // 🗎
+			columnSymbol.style.backgroundColor = "#9eb3c7";
+	
+			rowHeaders.append(columnSymbol);
+			for (let index in questionnaire)
+			{
+				//console.log(response[questionnaire][index]);
+				let columnHeaders = document.createElement("td");
+				columnHeaders.className = "questionnaireHeader";
+	
+				// Änderung der Header-Bezeichnungen
+				if (index == "name") columnHeaders.innerHTML = "Thema";
+				else if (index == "zeitstempel") columnHeaders.innerHTML = "Datum";
+				else if (index == "id") continue;
+				else if (index == "anzfragen") continue;
+				else if (index == "schueleranzahl") continue;
+				else if (index == "klassenname") columnHeaders.innerHTML = "Klasse";
+				else if (index == "fach") columnHeaders.innerHTML = "Fach";
+				else if (index == "bewertungsumme") columnHeaders.innerHTML = "Punkte";
+				else columnHeaders.innerHTML = index;
+	
+				columnHeaders.style.backgroundColor = "#9eb3c7";
+				columnHeaders.style.fontWeight = "bold";
+				columnHeaders.style.fontSize = "small";
+				columnHeaders.style.color = this.menuBarColor;
+				rowHeaders.appendChild(columnHeaders);
+					
+				let columnData = document.createElement("td");
+	
+				if (index == "zeitstempel")
+				{
+					let timestamp = questionnaire[index].split(" ");
+					columnData.innerHTML = timestamp[0];
+				}
+				else columnData.innerHTML = questionnaire[index];
+	
+				rowData.appendChild(columnData);			
+			}
+	
+			// Bogenstatus Header hinzufügen
+			let columnStatusHeader = document.createElement("td");
+			columnStatusHeader.className = "questionnaireHeader";
+			columnStatusHeader.style.color = this.menuBarColor;
+			columnStatusHeader.style.fontWeight = "bold";
+			columnStatusHeader.style.fontSize = "small";
+			columnStatusHeader.innerHTML = "Status";
+			columnStatusHeader.style.backgroundColor = "#9eb3c7";
+			rowHeaders.appendChild(columnStatusHeader);
+	
+			// Bogenstatus hinzufügen
+			let columnStatus = document.createElement("td");
+			columnStatus.style.fontWeight = "bold";
+	
+			let formDataCodes = new FormData();
+			formDataCodes.append("fbId", questionnaire.id);
+			var responseQuestionnaireCodes = this.Request("./php/main.php?mode=getCodes", formDataCodes);
+	
+			console.log("responseQuestionnaireCodes:");
+			console.log(responseQuestionnaireCodes);
+	
+			let codesArray = responseQuestionnaireCodes.split("},");
+			console.log(codesArray);
+	
+			if (codesArray != null && codesArray.length > 1)
+			{
+				columnStatus.innerHTML = "offen";
+				columnStatus.style.color = "#feb460"; // orange
+			}
+			else
+			{
+				columnStatus.innerHTML = "abgeschlossen";
+				columnStatus.style.color = "green";	
+			}
+	
+			rowData.appendChild(columnStatus);
+	
+			// TODO: Matthias eine Issue-Nachricht auf GitHub schreiben, die Methode getCodes muss optimiert werden
+			// Es darf nicht nur einfach eine Fehlermeldung kommen wenn es keine Codes für den Bogen gibt.
+	
+			table.appendChild(rowHeaders);
+			table.appendChild(rowData);
+			div.appendChild(table);
+			questionnaireList.appendChild(div);
+			
+			// Event-Listener zum Hinzufügen der Fragen bei einem Klick auf den Bogen (Öffnen des Fragebogens)
+			div.addEventListener("click", ()=>{
+	
+				this.ShowQuestions();
+				/*
+				for ( let i = 0; i < response.returnvalue[0].length; i++ )
+				{
+					// TODO: Iteration durch die Fragen um diese dem Bogen hinzuzufügen
+					// - Warten auf Anpassung der Backend-Methode "getFbFragen" (die Kategorien müssen noch hinzugefügt werden)	
+				}
+				*/
+			});
+		}
+
+
+	}
+
+	Request(path, formData)
 	{
-		this.questions.push(question);
+		let response;
+		let xhttp = new XMLHttpRequest();
+		xhttp.open("POST", path, false);
+		xhttp.send(formData);
+		response = xhttp.responseText;
+
+		//if (response.indexOf("{") != 0) return response;
+		//if (response.length > 100 && response != null) return JSON.parse(response);
+		return response;
 	}
-	
-	
+
+	ShowQuestions()
+	{
+		let formData = new FormData();
+		formData.append('fbId', div.id.toString());
+
+		var response = this.Request("./php/main.php?mode=getFbFragen", formData);
+		console.log(response);
+	}
 }
